# src/masterExpressApi.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "@bitgo/enclaved-bitgo-express",
    "version": "1.0.0",
    "description": "BitGo Enclaved Express - Secure enclave for BitGo signing operations with mTLS"
  },
  "paths": {
    "/api/{coin}/wallet/generate": {
      "post": {
        "summary": "Request to generate a new wallet",
        "operationId": "api.v1.generateWallet",
        "parameters": [
          {
            "name": "coin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user-agent",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  },
                  "enterprise": {
                    "type": "string"
                  }
                },
                "required": [
                  "label",
                  "enterprise"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletGenerateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "InternalErrorResponse": {
        "title": "InternalErrorResponse",
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "details"
        ]
      },
      "WalletGenerateResponse": {
        "title": "WalletGenerateResponse",
        "type": "object",
        "properties": {
          "wallet": {},
          "userKeychain": {},
          "backupKeychain": {},
          "bitgoKeychain": {}
        },
        "required": [
          "wallet",
          "userKeychain",
          "backupKeychain",
          "bitgoKeychain"
        ]
      },
...
    }
  }
}